fix: replace `post_update` hook with `target`
This makes it such that upstream users of `lean-cvc5` don't need
to run `lean update` before building.

This solution is a bit problematic, however, as it does not capture
that the `input_file`s depend on `cvc5Enums`. To avoid a potential
race condition, we do all of the work that the `cvc5Enums` job
is supposed to do in the "computing build jobs" phase.